# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pragati Bharati - Document Intelligence & Question Extraction Service",
    "description": "Scalable Document Processing & Question Extraction Service accepting PDFs and images containing examination material and converting them into structured, machine-readable questions with answer key association and confidence scoring.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/questions/review-queue": {
      "get": {
        "tags": [
          "Questions"
        ],
        "summary": "Get Global Review Queue",
        "description": "Retrieve questions across all documents that require human review\n(low confidence, missing options, cross-page split uncertainty, or unmatched answers).",
        "operationId": "get_global_review_queue_api_v1_questions_review_queue_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "is_reviewed",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by reviewed status",
              "default": false,
              "title": "Is Reviewed"
            },
            "description": "Filter by reviewed status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_QuestionReviewItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/questions/{question_id}": {
      "get": {
        "tags": [
          "Questions"
        ],
        "summary": "Get Question",
        "description": "Retrieve detailed information for an individual question, including its options and answer key.",
        "operationId": "get_question_api_v1_questions__question_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Questions"
        ],
        "summary": "Update Or Review Question",
        "description": "Human-in-the-loop review and correction endpoint.\nAllows reviewers to edit question text, options, question numbers, or approve questions.",
        "operationId": "update_or_review_question_api_v1_questions__question_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnswerKeyOut": {
        "properties": {
          "raw_answer": {
            "type": "string",
            "title": "Raw Answer",
            "description": "Raw answer text detected in source document"
          },
          "normalized_answer": {
            "type": "string",
            "title": "Normalized Answer",
            "description": "Standardized answer representation, e.g. 'A', 'True'"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation",
            "description": "Explanation or solution rationale if present"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Confidence score of answer detection",
            "default": 1.0
          },
          "status": {
            "$ref": "#/components/schemas/AnswerStatus",
            "description": "Status of the answer mapping",
            "default": "CONFIRMED"
          },
          "source_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Page",
            "description": "Page where the answer was found"
          },
          "association_method": {
            "type": "string",
            "title": "Association Method",
            "description": "Method used: INLINE_KEY_SECTION, SEPARATE_DOCUMENT, EMBEDDED, etc.",
            "default": "INLINE_KEY_SECTION"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "description": "Any warnings or uncertainties"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "source_document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Document Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "raw_answer",
          "normalized_answer",
          "id",
          "question_id",
          "created_at",
          "updated_at"
        ],
        "title": "AnswerKeyOut"
      },
      "AnswerStatus": {
        "type": "string",
        "enum": [
          "CONFIRMED",
          "INFERRED",
          "UNCERTAIN",
          "NOT_FOUND"
        ],
        "title": "AnswerStatus"
      },
...
      "ExtractionStatus": {
        "type": "string",
        "enum": [
          "CONFIDENT",
          "PARTIALLY_EXTRACTED",
          "NEEDS_REVIEW"
        ],
        "title": "ExtractionStatus"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OptionItem": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Option identifier, e.g. 'A', 'B', '1', 'i'"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Option text or statement"
          }
        },
        "type": "object",
        "required": [
          "key",
          "text"
        ],
        "title": "OptionItem"
      },
...
      "PaginatedResponse_QuestionReviewItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuestionReviewItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "PaginatedResponse[QuestionReviewItem]"
      },
      "QuestionOut": {
        "properties": {
          "question_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Number",
            "description": "Question number/identifier, e.g. '1', 'Q.2'"
          },
          "question_text": {
            "type": "string",
            "title": "Question Text",
            "description": "Full text of the question"
          },
          "question_type": {
            "$ref": "#/components/schemas/QuestionType",
            "description": "Categorized question type",
            "default": "MCQ"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/OptionItem"
            },
            "type": "array",
            "title": "Options",
            "description": "List of options for MCQ/Multi-select"
          },
          "source_pages": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Source Pages",
            "description": "Original page number(s) where question was found"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Extraction confidence score (0.0 - 1.0)",
            "default": 1.0
          },
          "status": {
            "$ref": "#/components/schemas/ExtractionStatus",
            "description": "Extraction status",
            "default": "CONFIDENT"
          },
          "review_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Review Reasons",
            "description": "Reasons flagged for human review"
          },
          "has_tables": {
            "type": "boolean",
            "title": "Has Tables",
            "description": "Whether question contains tabular data",
            "default": false
          },
          "has_images": {
            "type": "boolean",
            "title": "Has Images",
            "description": "Whether question contains embedded images/diagrams",
            "default": false
          },
          "metadata_info": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata Info",
            "description": "Additional structural metadata"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "is_reviewed": {
            "type": "boolean",
            "title": "Is Reviewed"
          },
          "reviewer_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Notes"
          },
          "answer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnswerKeyOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "question_text",
          "source_pages",
          "id",
          "document_id",
          "is_reviewed",
          "created_at",
          "updated_at"
        ],
        "title": "QuestionOut"
      },
      "QuestionReviewItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "question_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Number"
          },
          "question_text": {
            "type": "string",
            "title": "Question Text"
          },
          "question_type": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "status": {
            "$ref": "#/components/schemas/ExtractionStatus"
          },
          "review_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Review Reasons"
          },
          "source_pages": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Source Pages"
          },
          "has_unmatched_answer": {
            "type": "boolean",
            "title": "Has Unmatched Answer"
          },
          "is_reviewed": {
            "type": "boolean",
            "title": "Is Reviewed"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "question_number",
          "question_text",
          "question_type",
          "confidence",
          "status",
          "review_reasons",
          "source_pages",
          "has_unmatched_answer",
          "is_reviewed"
        ],
        "title": "QuestionReviewItem"
      },
      "QuestionType": {
        "type": "string",
        "enum": [
          "MCQ",
          "MULTI_SELECT",
          "TRUE_FALSE",
          "FILL_IN_THE_BLANK",
          "SHORT_ANSWER",
          "DESCRIPTIVE"
        ],
        "title": "QuestionType"
      },
      "QuestionUpdate": {
        "properties": {
          "question_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Number"
          },
          "question_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Text"
          },
          "question_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuestionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/OptionItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractionStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_reviewed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Reviewed"
          },
          "reviewer_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Notes"
          }
        },
        "type": "object",
        "title": "QuestionUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        }
      }
    }
  }
}